Move photon frontend internals in a dedicated util
